python/requirements/base.txt
aiohttp==3.5.4
dataclasses==0.5; python_version<"3.7"
pyyaml==3.13
async_generator==1.10; python_version<"3.7"
python-jose==3.0.1
python-dateutil==2.7.5
keyring==17.1.1
yarl==1.3.0
aiodocker>=0.14.0
click==7.0

-e .
